+++ flutter_auth/lib/auth_methods/TwilioOTPAuthMethod.dart
@@ -1,6 +1,7 @@
 import 'package:auth_provider/AuthMethod.dart';
 import 'package:auth_provider/AuthProviderUser.dart';
 import 'package:auth_provider/UserInterface.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:fly_networking/GraphQB/graph_qb.dart';
 import 'package:fly_networking/fly.dart';
 
@@ -8,34 +9,33 @@ class TwilioOTPAuthMethod implements AuthMethod {
   @override
   String serviceName = 'twilio_otp';
 
-  Fly _fly;
+  Fly fly;
   Node twilioLoginNode;
   String phoneNumber;
-  String language;
   String apiLink;
 
   TwilioOTPAuthMethod(
-      {this.twilioLoginNode, this.apiLink, this.phoneNumber, this.language}) {
-    _fly = Fly(this.apiLink);
-  }
+      {this.twilioLoginNode,
+      this.apiLink,
+      this.phoneNumber,
+      @required this.fly});
 
   @override
   Future<AuthUser> auth() async {
-    _fly.addHeaders({'lang': language});
-    Map result = await _fly.mutation([twilioLoginNode],
+    Map result = await fly.mutation([twilioLoginNode],
         parsers: {twilioLoginNode.name: AuthProviderUser()});
     AuthProviderUser user = result[twilioLoginNode.name];
     return user;
   }
 
   Future<void> sendSMS({String to}) async {
-    await _fly.mutation([
-      Node(name: "sendOTP", args: {"phone": phoneNumber, "type": '_' + to})
+    await fly.mutation([
+      Node(name: "sendOTP", args: {"phone": phoneNumber})
     ]);
   }
 
   @override
   Future<void> logout() async {
-    await _fly.mutation([Node(name: 'logout_user')]);
+    await fly.mutation([Node(name: 'logout_user')]);
   }
 }